requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
google-genai==0.8.0
python-dotenv==1.0.1
moviepy==1.0.3
librosa
numpy
supabase==2.10.0
pyjwt==2.10.0
httpx==0.28.0
